requirements.txt
Flask==3.1.1
mysql-connector-python==9.3.0
sshtunnel==0.4.0
paramiko==3.5.1
bcrypt==4.3.0
cryptography==45.0.5
PyNaCl==1.5.0

